#7 modify image uploader
