+++ web/uploaders/user_image_uploader.ex
@@ -1,3 +1,16 @@
 defmodule ForEctoUpgrade.UserImageUploader do
-  use ForEctoUpgrade.BaseUploader, "user"
+  use ForEctoUpgrade.BaseUploader, model: :user, field: :image
+  @versions [:medium, :small]
+
+  def transform(:medium, _) do
+    convert "100x100"
+  end
+
+  def transform(:small, _) do
+    convert "50x50"
+  end
+
+  defp convert(size) when is_binary(size) do
+    {:convert, "-thumbnail #{size}^ -gravity center -extent #{size} -format png", :png}
+  end
 end
